SOC-guidelines move file to previous place
