+++ src/test/java/com/wikia/webdriver/elements/mercury/components/discussions/desktop/Moderation.java
@@ -1,4 +1,4 @@
-package com.wikia.webdriver.elements.mercury.desktop;
+package com.wikia.webdriver.elements.mercury.components.discussions.desktop;
 
 import com.wikia.webdriver.pageobjectsfactory.pageobject.BasePageObject;
 import org.openqa.selenium.WebElement;
