+++ src/test/java/com/wikia/webdriver/elements/mercury/pages/discussions/PostsListPage.java
@@ -3,13 +3,13 @@ package com.wikia.webdriver.elements.mercury.pages.discussions;
 import com.wikia.webdriver.elements.mercury.components.discussions.common.ErrorMessages;
 import com.wikia.webdriver.elements.mercury.components.discussions.common.Post;
 import com.wikia.webdriver.elements.mercury.components.discussions.common.PostEditor;
-import com.wikia.webdriver.elements.mercury.desktop.BackButtons;
-import com.wikia.webdriver.elements.mercury.desktop.CommunityBadge;
-import com.wikia.webdriver.elements.mercury.desktop.HeroUnit;
-import com.wikia.webdriver.elements.mercury.desktop.Moderation;
-import com.wikia.webdriver.elements.mercury.desktop.PostsCreatorDesktop;
-import com.wikia.webdriver.elements.mercury.desktop.Promoting;
-import com.wikia.webdriver.elements.mercury.desktop.SortingTool;
+import com.wikia.webdriver.elements.mercury.components.discussions.desktop.BackButtons;
+import com.wikia.webdriver.elements.mercury.components.discussions.desktop.CommunityBadge;
+import com.wikia.webdriver.elements.mercury.components.discussions.desktop.HeroUnit;
+import com.wikia.webdriver.elements.mercury.components.discussions.desktop.Moderation;
+import com.wikia.webdriver.elements.mercury.components.discussions.desktop.PostsCreatorDesktop;
+import com.wikia.webdriver.elements.mercury.components.discussions.desktop.Promoting;
+import com.wikia.webdriver.elements.mercury.components.discussions.desktop.SortingTool;
 import com.wikia.webdriver.elements.mercury.components.discussions.mobile.DiscussionsHeader;
 import com.wikia.webdriver.elements.mercury.components.discussions.mobile.FiltersPopOver;
 import com.wikia.webdriver.elements.mercury.components.discussions.mobile.PostsCreatorMobile;
